First Phase/requirements.txt
branca==0.5.0
certifi==2022.9.14
charset-normalizer==2.1.1
cycler==0.11.0
dnspython==2.2.1
folium==0.12.1.post1
fonttools==4.37.2
geojson==2.5.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
Shapely==1.8.4
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.12
